Refactored to Julia Project structure.
