+++ Project.toml
@@ -1,0 +1,18 @@
+name = "SAMPLER"
+uuid = "4c8c32dc-9f2a-4d47-bd35-4132de4e6959"
+authors = ["Maximilian Schmeller <m.schmeller@tum.de>"]
+version = "1.0.0"
+
+[deps]
+ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
+Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+NetCDF = "30363a11-5582-574a-97bb-aa9a979735b9"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
+SeisSolXDMF = "fc3049fc-4624-4ea9-a8a9-22daae44174f"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[compat]
+julia = "^1.4"
