+++ src/main.jl
@@ -1,0 +1,2 @@
+import SAMPLER
+SAMPLER.main()
