+++ test/runtests.jl
@@ -1,0 +1,2 @@
+include("rasterization.jl")
+include("io.jl")
