--- test_resources/output-surface_cell.h5
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:93952c8b8b9bbf57591ba6ee030d9b4bfc205d25e63432a62a0c1719f54b99cb
-size 1670244692
