--- test_resources/output-surface_vertex.h5
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:1d916e43bad724e1e2301ac23581c20aa6859d83203237200b7c93d80a1eed2e
-size 2081424
